Applies server changes to local server
